[NuGet] Fix wrong FSharp.Core NuGet package restored
Creating a F# .NET Core console project, then installing FSharp.Core
4.5.0, would result in FSharp.Core 4.3.4 being used and displayed
in the Dependencies folder. FSharp.Core 4.3.4 is the NuGet package
implicitly added by the F# .NET Core SDK and was being used by
the project instead of the PackageReference in the project.
PackageReferences in the project will now override any implicitly
added NuGet packages which is the behaviour of dotnet restore on
the command line.

Fixes VSTS #636604 - Adding FSharp.Core 4.5.0 NuGet package restores
4.3.4 in .NET Core project
