+++ main/tests/test-projects/DotNetCoreFSharpCore45/DotNetCoreFSharpCore45.sln
@@ -1,0 +1,17 @@
+
+Microsoft Visual Studio Solution File, Format Version 12.00
+# Visual Studio 15
+Project("{f2a71f9b-5d33-465a-a702-920d77279786}") = "DotNetCoreFSharpCore45", "DotNetCoreFSharpCore45.fsproj", "{66FA1301-92BB-4721-AB4D-FA6BB685C3F1}"
+EndProject
+Global
+	GlobalSection(SolutionConfigurationPlatforms) = preSolution
+		Debug|Any CPU = Debug|Any CPU
+		Release|Any CPU = Release|Any CPU
+	EndGlobalSection
+	GlobalSection(ProjectConfigurationPlatforms) = postSolution
+		{66FA1301-92BB-4721-AB4D-FA6BB685C3F1}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{66FA1301-92BB-4721-AB4D-FA6BB685C3F1}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{66FA1301-92BB-4721-AB4D-FA6BB685C3F1}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{66FA1301-92BB-4721-AB4D-FA6BB685C3F1}.Release|Any CPU.Build.0 = Release|Any CPU
+	EndGlobalSection
+EndGlobal
